Use `std::string::c_str` instead of `std::string::front`
This code uses `front()` to get the underlying string buffer. However,
when glibc++ assertions are enabled, this causes an assert failure if
the string is empty. This failure is spurious in our use case because we
are just trying to get the string buffer; it's ok if it's "empty" (i.e.,
just the null byte). Instead, use `c_str()` which both captures our
intent better and avoids the assert failure when the string is empty.

Note that this assert failure only occurred with glibc++ assertions
enabled. Because Arch apparently enables them by default (while other
distros don't), it initially appeared to be an Arch-specific problem.
However, it's just that it only surfaced on Arch, while having the
potential for issues on other platforms.